+++ search.h
@@ -3,10 +3,11 @@
 
 #include "board.h"
 #include "stats.h"
+#include <stdint.h>
 
 /* Forwards to 'search' but just returns the move and does not expose the
  * full search record. */
-int searchm(Board *b, Player p);
+uint8_t searchm(Board *b, Player p);
 
 /* Returns a move (i.e. a column) selected by the KI engine. Any changes to the
  * board while searching are reverted before returning. */
@@ -14,7 +15,7 @@ void search(Board *b, Player p, SearchRecord *rec);
 
 void alphabeta_negamax(Board *b, Player p,
         float alpha, float beta,
-        int depth, int max_depth,
+        uint8_t depth, uint8_t max_depth,
         SearchRecord *rec);
 
 #endif	/* _SEARCH_H */
